# New getNode logic