```bash
$ go test -bench='BenchmarkFair/complex' -run=^$ -count=5 ./... | tee new.txt
...

$ benchstat old.txt new.txt
name                                   old time/op    new time/op     delta
Fair/complex-stable-flavor|complex-8     1.12ms ± 0%     0.87ms ± 0%   -22.40%  (p=0.008 n=5+5)
Fair/complex-chaotic-flavor|complex-8     277µs ± 1%      147µs ± 1%   -47.07%  (p=0.008 n=5+5)
Fair/get-stable-flavor|get-8              439ns ± 0%       41ns ± 0%   -90.75%  (p=0.008 n=5+5)
Fair/get-chaotic-flavor|get-8            5.06µs ± 0%     0.75µs ± 0%   -85.20%  (p=0.008 n=5+5)

name                                   old speed      new speed       delta
Fair/complex-stable-flavor|complex-8   1.15GB/s ± 0%   1.48GB/s ± 0%   +28.86%  (p=0.008 n=5+5)
Fair/complex-chaotic-flavor|complex-8   300MB/s ± 1%    566MB/s ± 1%   +88.93%  (p=0.008 n=5+5)
Fair/get-stable-flavor|get-8           1.47TB/s ± 0%  15.85TB/s ± 0%  +981.38%  (p=0.008 n=5+5)
```